+++ lib/progress.dart
@@ -2,4 +2,5 @@
 /// Consumers should accept a [ProgressState] stream.
 library progress;
 
+export 'src/progress/log_listener.dart';
 export 'src/progress/progress.dart';